Fix DJANGO startup.
