+++ codalab/lib/server_util.py
@@ -1,222 +1,13 @@
 """
 Utility functions used by the server applications.
+Don't import from non-REST API code, since this file imports bottle.
 """
-import binascii
+
 import base64
-import hashlib
-import hmac
-import random
-import struct
 import sys
-import time
 
 from bottle import request, HTTPResponse
 from oauthlib.common import to_unicode, bytes_type
-import six
-
-
-"""
-The following functions are adapted from the Django codebase and are:
-Copyright (c) Django Software Foundation and individual contributors.
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without modification,
-are permitted provided that the following conditions are met:
-
-    1. Redistributions of source code must retain the above copyright notice,
-       this list of conditions and the following disclaimer.
-
-    2. Redistributions in binary form must reproduce the above copyright
-       notice, this list of conditions and the following disclaimer in the
-       documentation and/or other materials provided with the distribution.
-
-    3. Neither the name of Django nor the names of its contributors may be used
-       to endorse or promote products derived from this software without
-       specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
-ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
-ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-"""
-
-# BEGIN ADAPTED FROM Django #
-
-# Use the system PRNG if possible
-try:
-    random = random.SystemRandom()
-    using_sysrandom = True
-except NotImplementedError:
-    import warnings
-    warnings.warn('A secure pseudo-random number generator is not available '
-                  'on your system. Falling back to Mersenne Twister.')
-    using_sysrandom = False
-
-
-def force_bytes(s, encoding='utf-8', errors='strict'):
-    """
-    Return a bytestring version of s.
-
-    :param s: string to coerce
-    :param encoding: encoding to use
-    :return: bytestring version of s
-    """
-    if isinstance(s, bytes):
-        if encoding == 'utf-8':
-            return s
-        else:
-            return s.decode('utf-8', errors).encode(encoding, errors)
-    else:
-        return s.encode(encoding, errors)
-
-
-def get_random_string(length=12,
-                      allowed_chars='abcdefghijklmnopqrstuvwxyz'
-                                    'ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789',
-                      secret=''):
-    """
-    Returns a securely generated random string.
-    The default length of 12 with the a-z, A-Z, 0-9 character set returns
-    a 71-bit value. log_2((26+26+10)^12) =~ 71 bits
-
-    :param length: length of random string to generate
-    :param allowed_chars: set of characters to use, as a string
-    :param secret: if system PRNG not available, cryptographically-strong secret
-    string to use in the random seed.
-    :return: generated random string
-    """
-    if not using_sysrandom:
-        # This is ugly, and a hack, but it makes things better than
-        # the alternative of predictability. This re-seeds the PRNG
-        # using a value that is hard for an attacker to predict, every
-        # time a random string is required. This may change the
-        # properties of the chosen random sequence slightly, but this
-        # is better than absolute predictability.
-        random.seed(
-            hashlib.sha256(
-                ("%s%s%s" % (
-                    random.getstate(),
-                    time.time(),
-                    secret)).encode('utf-8')
-            ).digest())
-    return ''.join(random.choice(allowed_chars) for i in range(length))
-
-
-def _bin_to_long(x):
-    """
-    Convert a binary string into a long integer
-    This is a clever optimization for fast xor vector math
-    """
-    return int(binascii.hexlify(x), 16)
-
-
-def _long_to_bin(x, hex_format_string):
-    """
-    Convert a long integer into a binary string.
-    hex_format_string is like "%020x" for padding 10 characters.
-    """
-    return binascii.unhexlify((hex_format_string % x).encode('ascii'))
-
-
-if hasattr(hmac, "compare_digest"):
-    # Prefer the stdlib implementation, when available.
-    def constant_time_compare(val1, val2):
-        return hmac.compare_digest(force_bytes(val1), force_bytes(val2))
-else:
-    def constant_time_compare(val1, val2):
-        """
-        Returns True if the two strings are equal, False otherwise.
-        The time taken is independent of the number of characters that match.
-        For the sake of simplicity, this function executes in constant time only
-        when the two strings have the same length. It short-circuits when they
-        have different lengths. Since Django only uses it to compare hashes of
-        known expected length, this is acceptable.
-        """
-        if len(val1) != len(val2):
-            return False
-        result = 0
-        if six.PY3 and isinstance(val1, bytes) and isinstance(val2, bytes):
-            for x, y in zip(val1, val2):
-                result |= x ^ y
-        else:
-            for x, y in zip(val1, val2):
-                result |= ord(x) ^ ord(y)
-        return result == 0
-
-
-if hasattr(hashlib, "pbkdf2_hmac"):
-    def pbkdf2(password, salt, iterations, dklen=0, digest=None):
-        """
-        Implements PBKDF2 with the same API as Django's existing
-        implementation, using the stdlib.
-        This is used in Python 2.7.8+ and 3.4+.
-        """
-        if digest is None:
-            digest = hashlib.sha256
-        if not dklen:
-            dklen = None
-        password = force_bytes(password)
-        salt = force_bytes(salt)
-        return hashlib.pbkdf2_hmac(
-            digest().name, password, salt, iterations, dklen)
-else:
-    def pbkdf2(password, salt, iterations, dklen=0, digest=None):
-        """
-        Implements PBKDF2 as defined in RFC 2898, section 5.2
-        HMAC+SHA256 is used as the default pseudo random function.
-        As of 2014, 100,000 iterations was the recommended default which took
-        100ms on a 2.7Ghz Intel i7 with an optimized implementation. This is
-        probably the bare minimum for security given 1000 iterations was
-        recommended in 2001. This code is very well optimized for CPython and
-        is about five times slower than OpenSSL's implementation. Look in
-        django.contrib.auth.hashers for the present default, it is lower than
-        the recommended 100,000 because of the performance difference between
-        this and an optimized implementation.
-        """
-        assert iterations > 0
-        if not digest:
-            digest = hashlib.sha256
-        password = force_bytes(password)
-        salt = force_bytes(salt)
-        hlen = digest().digest_size
-        if not dklen:
-            dklen = hlen
-        if dklen > (2 ** 32 - 1) * hlen:
-            raise OverflowError('dklen too big')
-        l = -(-dklen // hlen)
-        r = dklen - (l - 1) * hlen
-
-        hex_format_string = "%%0%ix" % (hlen * 2)
-
-        inner, outer = digest(), digest()
-        if len(password) > inner.block_size:
-            password = digest(password).digest()
-        password += b'\x00' * (inner.block_size - len(password))
-        inner.update(password.translate(hmac.trans_36))
-        outer.update(password.translate(hmac.trans_5C))
-
-        def F(i):
-            u = salt + struct.pack(b'>I', i)
-            result = 0
-            for j in range(int(iterations)):
-                dig1, dig2 = inner.copy(), outer.copy()
-                dig1.update(u)
-                dig2.update(dig1.digest())
-                u = dig2.digest()
-                result ^= _bin_to_long(u)
-            return _long_to_bin(result, hex_format_string)
-
-        T = [F(x) for x in range(1, l)]
-        return b''.join(T) + F(l)[:r]
-
-
-# END ADAPTED FROM Django #
 
 """
 The following functions are adapted from flask_oauthlib.utils and are
