+++ codalab/objects/user.py
@@ -7,7 +7,7 @@ import hmac
 
 from codalab.common import UsageError
 from codalab.model.orm_object import ORMObject
-from codalab.lib.server_util import force_bytes, get_random_string, pbkdf2, constant_time_compare
+from codalab.lib.crypt_util import force_bytes, get_random_string, pbkdf2, constant_time_compare
 
 
 class User(ORMObject):
